[Shuwen&Songbai] render radar node.

+++ ARTechRadar/Service/RadarDataService.swift
@@ -38,8 +38,8 @@ class RadarDataService {
     }
     
     private func parsePosition(postion: [String: Any]) -> Vector2D {
-        let x = postion["X"] as! Double
-        let z = postion["Z"] as! Double
+        let x = postion["X"] as! Float
+        let z = postion["Z"] as! Float
         return Vector2D(x: x, z: z)
     }
 }
